feat: добавлена кодогенерация сериализации компонентов на сцене. прокинута сериализация дефотных типов
