+++ directEngin/lib/serialization/base.h
@@ -1,0 +1,12 @@
+#ifndef  __SERIALIZATION_FORWARD__
+#define _SERIALIZATION_FORWARD_
+
+//@brief mark Type as Serializable. mean you builder will create a common serializer.
+#define SERIALIZE
+
+//@brief mark type file  as serializable. mean you builder will add field to common serialization
+#define SERIALIZE_FIELD
+
+
+
+#endif //! __SERIALIZATION_FORWARD__
